wip(resources): cleanup resources and make custom resource the default resource instead
